Move quantized bmm metadata workaround to ArmPass
